requirements.txt
Django==4.0.3
djangorestframework==3.14.0
PyJWT==1.7.1
psycopg2==2.9.3